Switch to signing decorator
